+++ Modules/Intent.Modules.Blazor.Components.MudBlazor/release-notes.md
@@ -4,3 +4,4 @@
 - Improvement : Support for sub menu items.
 - Improvement : Support for `Layout` attributes on `Containers`.
 - Improvement : Support for `OnSelected` event on `Select`.
+- Improvement : Support for `Radio Button Group`